requirements.txt
Flask==3.0.0
pymongo
Werkzeug==3.0.1
pytest==7.4.3
gunicorn==21.2.0
python-dotenv==1.0.0
